Clean-up
* Remove unused code
* Use generics
* Various small clean-ups

+++ BuiltInChips/src/main/java/builtInChips/DMux.java
@@ -28,7 +28,7 @@ public class DMux extends BuiltInGate {
     protected void reCompute() {
         short in = inputPins[0].get();
         short sel = inputPins[1].get();
-        outputPins[0].set((short)(sel == 0 ? in : 0));
-        outputPins[1].set((short)(sel == 0 ? 0  : in));
+        outputPins[0].set(sel == 0 ? in : 0);
+        outputPins[1].set(sel == 0 ? 0 : in);
     }
 }
