+++ BuiltInChips/src/main/java/builtInChips/DMux4Way.java
@@ -29,9 +29,9 @@ public class DMux4Way extends BuiltInGate {
     protected void reCompute() {
         short in = inputPins[0].get();
         short sel = inputPins[1].get();
-        outputPins[0].set((short)(sel == 0 ? in : 0));
-        outputPins[1].set((short)(sel == 1 ? in : 0));
-        outputPins[2].set((short)(sel == 2 ? in : 0));
-        outputPins[3].set((short)(sel == 3 ? in : 0));
+        outputPins[0].set(sel == 0 ? in : 0);
+        outputPins[1].set(sel == 1 ? in : 0);
+        outputPins[2].set(sel == 2 ? in : 0);
+        outputPins[3].set(sel == 3 ? in : 0);
     }
 }
