+++ BuiltInChips/src/main/java/builtInChips/DMux8Way.java
@@ -29,13 +29,13 @@ public class DMux8Way extends BuiltInGate {
     protected void reCompute() {
         short in = inputPins[0].get();
         short sel = inputPins[1].get();
-        outputPins[0].set((short)(sel == 0 ? in : 0));
-        outputPins[1].set((short)(sel == 1 ? in : 0));
-        outputPins[2].set((short)(sel == 2 ? in : 0));
-        outputPins[3].set((short)(sel == 3 ? in : 0));
-        outputPins[4].set((short)(sel == 4 ? in : 0));
-        outputPins[5].set((short)(sel == 5 ? in : 0));
-        outputPins[6].set((short)(sel == 6 ? in : 0));
-        outputPins[7].set((short)(sel == 7 ? in : 0));
+        outputPins[0].set(sel == 0 ? in : 0);
+        outputPins[1].set(sel == 1 ? in : 0);
+        outputPins[2].set(sel == 2 ? in : 0);
+        outputPins[3].set(sel == 3 ? in : 0);
+        outputPins[4].set(sel == 4 ? in : 0);
+        outputPins[5].set(sel == 5 ? in : 0);
+        outputPins[6].set(sel == 6 ? in : 0);
+        outputPins[7].set(sel == 7 ? in : 0);
     }
 }
